pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "peergrading"
version = "0.1.0"
description = "Peer grading pipeline starter"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.1",
  "openpyxl>=3.1",
  "PyYAML>=6.0",
  "python-dateutil>=2.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
